ref(replay): Extract performanceObserver out
